[refactor] 섹션 스크롤 useSectionScroll로 리팩토링
